Cleanups related to FMan(Item) setup
In the recycler view use the path property/member instead of the
template's member named content (which was renamed to path; did
this rename manually, if I had used the refactor option it would
have updated this automatically).

Convert from path string to path object using Paths class.

Uses Files.isDirectory to check if the current directory entry is
a directory or a file.

I had to switch to minimum API level of 26 to support above. Later
may change back to a lower minimum API level by updating the logic
as required.

	May have to add some custom helper routines.