Main activity call singleton FMan loadPath
As a android app cant directly access the root folder in non-rooted
mode, so cant directly try to load dir entries of system-root i.e /
dir by default.

I have to get the application related directory.

As FMan is a singleton, which is potentially implicitly created
(I assume, I have to look into kotlin singletons still), so at
that time the android context is not directly available.

So for now have moved the logic to load the default dir entries
into the actually appropriate loadPath helper function. And inturn
the same is called from the main activity, which has access to the
android context and inturn the getFilesDir/...

NOTE: One is required to call FMan.loadPath atleast once with a
valid non-null path string.