MainActivity: Fix layout
Had forgotten to set the orientation of the main linearlayout, that
was the thing which was creating the fragment view not showing on
screen issue after the button+textview linear layout was added at
the top. Have fix this to be vertical as required now.

Have also embedded the fragment within another linear layout set
to vertical for now. May later change to horizontal.

Instead of directly embedding a string in the layout file, have
changed to use a string resource value for the same.