FMan: Always maintain curPath as absolute
backPath logic was not converting the new path as a absolute path,
as the current directory relative to which the path will be taken
is not something one can be sure of, so updated the logic to store
curPath always as absolute path.

Also enabled the log message of MainActivities backPath.

TODO: The screen contents are not updating sometimes, not sure if
it is some bug in android(x) framework or some bug in the emulator
used for testing or some oversight in my flow.