Main: Check WriteExtStorage, Request, StillNo Quit
Manifest updated to request WriteExternalStorage permission.

However in new android, this wont be granted by default, so we have
to check if its already granted or not, if not, then request user
to grant the permission.

If user doesnt grant the permission, then stop, as we cant do without
this.