RequestLegacyExternalStorage OnPath2 ManageExternalStorage
For now use requestLegacyExternalStorage to get access to the
root of the external storages.

TODO:

As Android 11 onwards is moving into scoped storage, READ/WRITE
ExternalStorage will no longer be useful/sufficient. Rather one
will have to use the new ManageExternalStorage permission, which
will be allowed for a select group of applications like filemanagers,
backup/restore apps, etc only.

As this is a filemanager+share+backup/restore kind of app, so it
should be fine.