getVolumes cleanup
Return a ArrayList instead of a Array, it should be more efficient
while adding things to the array internally.

Added a getBasePath helper which extracts the base path from a given
path, by checking if it contains a marker specified. For android's
app related paths, Android is the marker.

i.e from /crazyandroid/mount/path/Android/data/what/else
get /crazyandroid/mount/path

The same is used to extract the base path for browsing.

Moved the logging to be next to where the corresponding path is got
from the system.

Use getExternalFilesDirs to get all external storage devices related
info from app perspective, instead of the primary external storage
got using getExternalFilesDir.

NOTE: In current android versions this also means that there is no
need for one to potentially use StorageManager to get the storage
volumes available in the system and then inturn use the hidden func
to try and extract its path.