Main: Just to show, Handle dialog onclick in Main
Here instead of the previous runtime temporary class instance based
handling of the AlertDialogs list selection, the MainActivity class
is extended to implement the DialogInterface.OnClickListener interface
and inturn the same is used to handle the VolumeSelector alertdialogs
list selection.

This is useful, if one wants to handle multiple dialogs from a single
handler in a similar way.

However as we dont want this, so I am showing this here, just for my
own purpose, as a reminder. I will be reverting this next.