FMan: id is a int, also stored in view holder
The id also indirectly identifies the position in the FMan's item
list, because currently the position and id match one to one.

NOTE: However a Recycler view associated with a list view will only
contain a subset of items, so a position in the recycler's list of
view holders doesnt match the position of the corresponding entity
in the FMan's item list, so having this explicit id helps.