FManFrag:P2: RecyclerView VHolder.pathView onClick
As previous trapping of recycler view's onClickListner doesnt
seem to have helped, so now I go even lower in the view hierarchy
and directly trap the pathView gui widget in the ViewHolder and
that seems to have helped. Because now I get the log message wrt
the item, which I am logging from this pathView's onClickListner
handler.

NOTE: as I am only trapping pathView's onClickListner, so if one
clicks on typeView, it doesnt trigger the handler. ie it's behaving
like what was expected.