FManFrag:P3: RecyclerView onClick of ViewHolder
Now instead of setting the pathView's onClickListner in the ViewHolder,
I set it for the ViewHolder's full view (which is inflated in Adapters
onCreateViewHolder and passed to ViewHolder).

So now I can either click either the pathView or the typeView and in
both cases, the onClickListner is triggered, as expected.

TODO: Have to either remove that RecyclerView onClickListner and or dig
that path bit more.