FMan ItemSelection from ViewHolder to MainActivity
In FMan added a Interface for handshaking FMan Item selection event,
as well as a placeholder to hold a reference to its implementation.

MainActivity provides a implementation of this interface and inturn
updates the corresponding reference for the same in FMan.

FManRecyclerViewAdapter ViewHolder's onClickListener updated to call
any implementer of FManItemSelectIF's onSelectListener through FMan's
reference to the same.