FMan: Dir Entry type a enum with short string desc
The ITEMS of FMan singleton i.e the Dir Entries, had a type member
which was previously defined as a string, so one could try assign
any value into it.

So now it has been changed into a newly defined enum, so that it
can be only assigned to the enum values defined for it.

However the original by-convention string representation was also
used to display the same to user, which inturn was only a single
char string. But naming the enum values has a single char makes
them less implicitly identifiable, so rather the enum names/values
have been kept to be longer and self explanatory, while at same
time a member property called shortDesc is defined for this enum
class, so that inturn each of the corresponding enum values can
have this assigned to the original single char string.

    THis ensures that in most places the enum values are directly
    used, and inturn only in the place were it was used to show
    type to user, now the shortDesc corresponding to the enum value
    is shown. In future, if we want to change the associated short
    desc and or add new types, it can be done at a single location
    i.e the enum class definiation for the FManItemType.