Lint: Avoid redundancy type info
When assigning a value to a val/var during its decleration there
is no need to explicitly specify the type, as it can be infered
from the value most of the times.

No need to convert a object explicitly to string within in string
template (used for printing/logging/...)

The type of function/interface expected as a argument by a function
is already known, so no need to explicitly specify it again for a
anonymous class/function.