FManFrag: Saner way of handling onClick on RecyclerView
However its of no use, as there is no direct identifier as to which
viewholder/item was selected from the list. Rather the viewholder->
onClick logic which I have already implemented seems to be the simpler
and better option compared to adding lot more complexity and classes
with recyclerview-selection.

But otherwise this code shouldnt crash, as it tries to do things in
a more saner manner.