FManFrag:RecyclerView: Start at 1st item at beginning
Logic fixed such that the 1s time one uses the keyboard down logic
it selects the 1st item in the list.

However as one goes about using GUI UP btn to switch to a different
list, the listIndex is not reset, so it will continue from where one
left in the last list ie wrt position.

TODO: As I got beyond the last item in the list, I have to check, where
the keyboard focus shifts, something seems odd at that situation.
However for a initial experiment it is fine.