RecyclerView: Fix oversight wrt listIndex, minimal log
Previously the logic was trying to go 1 item beyond what is there
in the item list, this has been fixed now.

Now log only when one is acting on the key press proactively.

TODO: Noticied that the getChildAt doesnt return the view which
maps to actual list, but more of some internal list maintained by
recycler view.