Main:FMan RecyclerView DPAD Center press loadPath if ok
If the user selects a directory using the keypad, then load the same.

The logic uses the already implemented FManItemSelectIF logic of Main
to achieve this.

NOTE: As their could be directories which dont have any content in them,
so we require to take care of this possibility in FManItemSelectIF logic.
THe logic has been updated for the same. TODO: Have to take care of -ve
index which could also be set initialy.