RecyclerView single item fix
If there is only a single item in the list, the previous flow of
highlighting the cur item and dehighlighting the old item would
lead to the highlight getting dehighlighted.

So have changed the flow to fix the above issue.

TODO: However if there are more items than can fit the screen, then
as getChildAt seems to return a viewholder based on some internal
ordering of list of viewholders (may be latest used order or ...)
and its position and not the logical item's position, so the item
highlighting goes for a toss, this needs to be fixed.