RecyclerView: cycle through items with runnable
Using scrollToPosition instead of smoothScrollToPosition also didnt
help with the race that seemed to be there between calling this and
calling findViewByPosition.

So logically it appears like RecyclerViews' scrollToPosition, doesnt
process anything immidiatly (not even updating its viewholder list),
instead it seems to post some message/event into the event-run-loop
of the gui logic, which will be acted on after the current event
handler which is calling scroll and findviewby etc returns.

So findViewByPosition will either not have the view for that given
position (because the view is currently not visible, and will become
visible only after the scroll). And Or potentially could have a view
that will get changed (less likely, but one never knows, as one doesnt
have control over that code), which can lead to issues if we are
changing that view in any way.

SO now the logic to findViewByPosition and inturn operate on that
view has been moved into a runnable which gets posted to the UI
run loop. This should ensure that this will get called only after
the scroll related message has been handled (bcas it got posted
before we posted our runnable). And thus we should be operating
on the right view.

NOTE: If required one could operate on the prevIndex related view
directly in the call which changes/scrolls/cycles through the list
entities, because it is already visible when this call is being
processed. However as we should be in a situation where more than
1 or normally atleast few items are always visible in the GUI, so
handling both prevIndex and newIndex of the list in the runnable
should be fine for now.

This runnable based logic seems to have fixed the ui updation of
selection / active item coloring related issue for now, as was
logically expected.