RecycleView: View Color changing func, use going into dir
Moved the logic which gets the view associated with a position and
then set its background color, into a new helper function.

In turn the same is used for updating item view highlighting when
scrolling/navigating using keypad.

Now additionally the same is called to unhighlight a potentially
already highlighted view, when navigating into a subdir.

TODO: Same needs to be done when navigating into the parent.