Main:Fman:RecyclerView highlight clear in loadpath
Added a new clearHighlights function to FManFragment.

To allow any highlighting in the recycler list to be cleared when
going up into the parent directory, now the loadPath of Main calls
this new helper function, before it resets the listIndex.

TOTHINK: may be merge both clearing highlighting and reseting listIndex
into a single function.

Also the last commit logic to clear highlighting when navigating
into a child dir, had a corner case in terms of I was calling the
clear logic, when ever enter/dpad_center was pressed irrespective
of whether the current item is a dir or a file.

So now I no longer call the clearHighlights in the key event handler.
Instead the updated loadPath of Main which either way gets called
for the key event also, takes care of doing the dehighlighting job.
This also has the added advantage that it triggers only for directory.

TODO: There could be a corner case that if we dont have rights to go
into the dir that the user has selected, then this highlight clearing
as well as listIndex reseting wont get triggered. While the logic will
reset the overall view to be one of the user selected main storage
base paths. In this case as loadPath will again get called, so it should
sort things up wrt clearing and index. But as we show the topmost
base path contents by default if a failure occuring in loadPath, while
parallely waiting for user to select from the list of base paths in
the system, will it create any corner case or not needs to be thought
through once. Seems less likely as the highlighted item's index is
still retained till a valid load occurs at which time both clearing
of highlighting as well as index reseting occurs.

TODO: NOTE: As the color used for highlighting and reseting back to
a unselected state is hardcoded in the program, it may not play
properly with themes which force different colors in the system
wrt the recycler view. There would have been a simple way of resolving
this had there been a way to get the existing background color of a
text view before one changes it, so that it can be restored later.
But the textview doesnt seem to provide a mechanism for this. Need
to look at the TextView and View/GUI in general documentations of
android once.