Make Path entry take full width of display beyond type
Previously the text view corresponding to the path info was only
occupying the space required for its content(i.e text string), so
one could not use touch or mouse press beyond the text to trigger
a navigation into a directory.

So now have changed the GUI config xml so that the view holder
view and inturn the text view corresponding to the path match the
width of the parent as much as possible (obviously after leaving
space for it's sibling ui elements). With this now the user can
click any where along the width of the display. Also the keypad
based navigation focus highlighter will highlight the full width.