FManFrag: Add a highlightRecyclerItem helper
This saves the existing background color, before changing it to
highlight the item. And inturn if highlight is passed as false,
then it restores the saved background color back for the item.

NOTE: It maintains only a single global saved background color,
so it cant maintain unique background colors for more than 1 item,
if they are different, as the last highlight call's value is what
will remain.