FManFrag: RecycItem Highlighting update
Found that some times the background drawable is not available
sometimes and sometimes it seems to be wrong (like a different
items background drawable has been assigned to this and not yet
updated for this item kind of thing). So this was leading to
crash, which was worked around, but then the 2nd issue of drawable
in transition was potentially making the highlighted color appear
at different places as the background color (atleast ie what the
end result appeared as).

So have changed the logic to use the backgroundTintList if available
else fall back to white. I seem to be getting fallback white mostly
as of now (based on end result, havent checked by putting debug
print in code yet).

But otherwise the logic seems to be doing the job.

Have replaced previous colorRecyclerItem with highlightRecyclerItem
calls now. Also as background saving occurs only on calling with
highlight true first, so have changed the sequence of calls by
setting cur item's highlight first and only then clearing the prev
items hightlight. The idea being the saved colorBackground will
be from the viewholder, even the first time clear highlight is
triggered. NOTE: If I check if the prevIndex is valid or not, I
can avoid this new sequence and go back to old sequence and inturn
match save and restore at a individual item level, but as all
items in the list should have same color scheme so for now shouldnt
matter.