RecyclerView:2Revert: Try avoid view state outofsync
Have tried the things in this commit to try and avoid the odd
display view highlight non clearing issue when going into a new
directory, due to a previous highlight not getting cleared.

But none of these seem to be helping. This is just to capture
at one place. Will be reverting this commit next.

INFO: The issue is that if one goes into a directory which has
lot more items than can fit in the screen and then loop through
it more than once without selecting any of the directory. And
in the 2nd or 3rd run through the items, come to a item which
is outside the initial view by scrolling down to it, and then
if one enters this directory, then the new updated list view
created with contents of this directory will contain one of its
items as being highlighted already and wrongly (this is only a
issue with the view having not got cleared, otherwise indexList
is actually proper).

THis also contained the scroll to begining of the list fix, which
I will apply later again, after the revert of other changes.

INFO: If one scrolls throu a directory content which cant fit in
the screen. And inturn enter one of its subdirectories which is
not visible till one scrolls few steps down, and if this subdir
also contains lot of entries, then it wont be showing the list
from the 1st item, rather it will be already scrolled down a bit.