RecyclerView: Oh RecyclerView, strange ways or too good
What I had logically assumed with caution to check later, but given
that logically it should be fine in most probability, has turned out
to be not satisfied by the RecyclerView implementation and or one
possible corner case in that normal flow is striking back.

The assumption noted had to do with earlier today when adding the
runnable wrt highlighting of current item as part of keypad based
navigation.

When scrolling down (and only down) and if more than 1 item can fit
the screen, then ideally even if I delay (so that the scroll down can
occur and the item viewholders updated suitably) the clearing of the
previous item's highlighting so that it is done along with highlighting
of the current item, it should be fine, as prev should still be in view
and cur should have now come into view. However as noted in last 2 or 3
commits, I was finding that in some particular sequence of interactions
involving long lists which dont fit the screen, it was not clearing the
previous item's highlight some times, but other times it was doing it
fine. The only difference being how many times I loop throught the full
list again and again before entering into a subdir with lot of contents
again.

So have changed the flow to clear the prev item immidiately, while letting
the highlighting of the current item to remain in the runnable for later,
so that scroll can do its magic of getting it into view. This inturn
seems to have fixed the issue of inconsistancy noted above.

THe one possible corner case mentioned above is that as the logic is
implemented to allow loopback to begining when the end is reached, at
that instance the prev item is no longer in view, so the 100% safe
assumption cant apply. However when there is no pressure, ideally
it shouldn't follow a different path only sometimes, but either way
the restructuring flow to be double safe seems to have done the trick.

Also I have added the scrollTo 1st item logic now, when ever a new
content list is being shown.