Main: Kotlin cant set intent.data? NewAndroid on file uri intent
The problem was that one cant directly assign the data field of
the intent by assigning to it, it seems. Even thou kotlin doesnt
raise any error.

Now either creating a intent with the uri passed to constructor and
or by using setDataAndType seems to work. (i.e Kotlin issue resolved)

However android doesnt allow file:/// scheme based uri's to be
passed from one application to another and inturn will kill the
app, if it does it.

    One could create a empty VmPolicy for the StrictMode and bypass
    this check.

    Or one could try sharing using intent with send action and
    grant read uri permission flag.

    But both of the above may not work beyond Android 9.

    One is required to potentially implement the FileProvider
    mechanism and inturn grant read uri permission flag to get
    what I am trying to do here, on Android 10 and above.