FlexiNav: rm useless RecyclerView.onKeyDown, cleanup msg
Have removed the uselss onKeyDown extension function, bcas it doesnt
get called directly and or normally, if some other implementation is
already there, for sure. Even when nothing else handles keypresses,
still it doesnt seem to get called.

Maybe if one extends the class, it can be used. But extending the
RecyclerView itself wrt a custom onKeyDown handler doesnt seem to
help.

Also have updated the onLongClickHandler log to mark it has a long
click.

NOTE: Wrt AndroidAPI29 and RecyclerView it is found that both Android
API29 normal ones (mobile/tablet/...) as well as the TV one seem to
now handle the  keyboard and dpad respectively and inturn convert
them into mouse/touch clicks. And in this case, the long press to
long click works for dpad, but not for keyboard input.

The reason this doesnt work for keyboard input is because keyboard
firmwares/drivers normally convert a long keypress into multiple
independent keypress repeat events. So even thou one gets one long
press -> long click event, but one also gets a bunch of invalid
click events following it which creates a problem.

However if we remove keyboard from the equation, then wrt mouse/touch
and dpads.

On AndroidAPI25 and before if one provides mouse/touch event handlers
of recyclerview, then that in itself is not sufficient for a TV
environment, where one will have to handle the remote/dpad events,
by using onKeyListener and DPAD keycodes.

On AndroidAPI29 and later if one provides mouse/touch event handlers
of recyclerview, then it automatically handles both mouse/touch
interactions as well as dpad/tvremote interactions.

TOCHECK: However one thing that was noticied is that, AndroidAPI29
on Mobile provides visual indication of where the focus is when
using keyboard. But AndroidAPI29 on TV (which is still in preview)
doesnt provide the visual indication of where the focus is when using
dpad/tvremote.

For android versions between 25 and 29, where they stand wrt RecycView
and dpads/tvremtoes needs to be cross checked.

UPDATE: Android API 28 also showing same behaviour as API 29.