SimplifiedListTheme:P3: Use activated instead of focus
As focus didnt seem to change so use activated for now.

NOTE: As I was able to jump from anywhere in the list (ie after
any number of down key presses) to the button on top by just
pressing the up key once, so had a feel that focus is not moving
into the individual list items. This is confirmed now, from a
different angle.

Need to read docs to see which function/mechanism should be
used to get focus into items of the list and inturn change
between them as required.