RecyclerView: Trapping and Using up DPadUP key didnt help
Wanted to check, if just in case, not consuming DPap UP key was
creating any problem. THis was less likely, as even a single dpad
up was taking focus back to UpBtn. But still wanted to cross check
once and its confirmed that even not consuming dpad up key was not
the problem.