RecyclerView:P5:Focusable in layout, key/dpad auto handling
** Prelogue
When trying to handle focus with manual key event handling

** Reality hits
After all the circus, found that Rather in Older Android API levels
of 25 and below, the recyclerView items by default are not focusable,
so one has to explicit set it has focusable in the layout xml file.

WHile in API levels 28 and above, the recyclerView items are setup to
be focusable, by default, without requiring to do any additional entry
for focusable in the layout xml file.

In turn if the items are focusable, then key events are handled by
android automatically to allow one to navigate thro the recyclerView
items.

This is not clear from documentation, and especially, If one starts
from Android API29 where keys are handled automatically and moves to
API25 where it is not working, it will confuse people. Ideally this
change in behaviour should be mentioned in the document, so that one
doesnt have to waste time trying to see why the behaviour has changed
and inturn try to implement the key handling logic on their own.

NOTE: However with this now, even if one implements a onKeyListener
in their recyclerView, as it stands now in android, it no longer
comes/traps into it. So it would appear like one will have to create
a child class of recyclerView to try and control it. THe reason this
may be required is, like if one wants to use say right key to jump
to the top from a vertical list quickly.