RecyclerView:TrapKeyEvents: ViewHolder setOnKeyListener
If one wants to trap the key events when working with a recyclerView
then use it's ViewHolder's setOnKeyListener. And as usual, if you
handle the key event fully, then return true, else return false so
that additional logic provided by android/widget can do its job.

However if one wants to handle the key events at a higher level in
the GUI heirarchy like at the full recyclerView level and not at the
individual item level, then setting onKeyListener at the recyclerView
level doesnt do the job (if focus is enabled for the items, else it
will work partly)