FManFrag: updateFrag scroll if position given
updateFrag function updated to take a optional position argument.
If it is non negative, then scroll to it.

PReviously it was always scrolling to the 1st item in the list.
Now instead it scrolls to the given position.

THis will be useful, when we want to go back up the directory
hierarchy.