Main:back:P1: FMan indexOf path, loadPath position
FMan indexOf searches for the given path and returns the position
of it in the list of FMan item entries.

loadPath takes a defEntry argument, which if given and inturn found,
in the given path, will be used to position the directory entries
list to be at that given entry.

backPath passes the path it is backing up from to loadPath, so that
the corresponding entry can be positioned in the list of entries.