Directory entries selection basic skeleton
Have added a handleSelect helper func to FManRecyclerViewAdapter,
it takes care of adding or removing the specified item from the
selected list. As well as it calls any doSelect call back.

Currently selected list is implemented as a ArrayList of integers.
TODO: Later may change to ArrayList of Paths, if allowing items
to be selected from across directories.