RecyclerView Items selected list - P2
Previously I was activating or deactivating only when binding the
view holder to data of item at a given position.

Now when a item is being selected or unselected immidiately the
corresponding view is activated or deactivated. Otherwise the
activation may not be visible as the binding will only occur during
the initial creation of the list view and or later when scrolling
over a long list of items, and we require to reuse the limited views
used to show the list by recyclerview.

    So a small list will trigger a rebinding after the initial binding
    so the change in activation state will not get updated.

    Now this update to logic takes care of this case.

NOTE: The FManRecyclerViewAdapter takes care of maintaining the
selected list of items and inturn manage the visual indication wrt
the same as required. At same time doSelect of FmanItemInteractionIF
is called so that user of FMan and its UI can handle any additional
logic if it requires.