FManDataP4: Brute conversion of MainActivity plus
Have converted Main activity and associated logic to work with
FManData mainly from a compilation perspective for now, in a brute
force / some what blind manner.

THe flow and structures may require tweaking to better handle corner
cases if required. Need to think about this once and update if reqd.

THis is a quick and dirty update for now. And in normal case it is
working for now.