setupStartState including Loading path only once
When the program is starting out, with the new FManData and related
restructuring based flow

* the path was loading twice

This was fixable by avoiding the loadPath from MainActivity (bcas
FManFrag already does it), and inturn setting the current path
related text view.

* if the activity was restarted due to returning from a activity
it triggered or because of device orientation change or so, then
the saved path in savedInstance was not getting passed to the
FManFrag's initial loadPath, so also necessitating the additional
loadPath in onCreate of MainActivity, which is mentioned above.

Now have added a setupStartState logic which gets called before the
MainActivity inflates its view, which also triggers the FManFrag.

This inturn checks if there is any savedInstance based path to load,
if so the same is passed to FManFrag, so that it can use the same,
when it is starting out. THis also helps avoid the additional loadPath
in MainActivity's onCreate, as already mentioned.

TODO: Use this setupStartState to check if we are on a TV or on a
mobile or tablet in landscape mode, in which case, we switch the
FManFrag to use a 2 column grid layout rather than a linearlayout
for the recycler view based file listing.