setupStartState: Orientation check before TV check
Check for TV after checking for Landscape|Orientation so that TV
check overrides results of orientation check.

THis ensures that if we are on TV, then irrespective of portrait
or landscape orientation, it will setup 2 column grid for the file
listing. As TV screen is physically very large, so 2 column grid
is fine even for portrait mode at one level, so this flow.