WebView: Enable zoom control, try zoom out
Last commit involving loadWithOverviewMode and useWideViewPort allowed
fitting the width to the screen suitably. However if the aspect ratio
is such that height is very large, then the webview logic seems to allow
width to reduce upto ~50% of screenwidth and after that it stops from
reducing the image further to fit it on the screen.

Even enabling zoom control on screen to allow user to change it also
doesnt allow zooming out beyond this.

Also the zoomOut calls dont seem to help.

Also the brute force and race happy post (in this case as it just
gets delayed by some amount and not till the rendering is over, at
one level), doesnt get the proper contentHeight.