WebView: Similar to zoomOut, zoomBy also doesnt seem to help
At one level to be expected as both would be using the same underlying
logic.

NOTE: This could also be because, we have set overviewmode and wideport
to true. UPDATE: Even avoiding these two still doesnt help wrt zoom.
zoom doesnt seem to work for images. However again independent of overview
and wideport once for a text file, it zoomed. UNLESS android install into
avd goofs up when the apk being tested is previously installed and a new
update is being done to it.