ViewFile: IntThenExt, ExtThenInt logics
** Try InternalViewer only if required ExternalViewer

This is the default flow, here first we try to view the file
using internal viewer, if a exception occurs or if InternalViewer
returns a error, then the external viewer is tried.

** Try ExternalViewer and if possible and required InternalViewer

THis is a alternate flow provided, which is disabled by default.
Here the External Viewer is tried first. However if there is a
exception then the Internal Viewer is tried.

As the ExternalViewers triggered on a default android installation
from google (in their avd's) where found not to return consistent
result code wrt success or failure. In this case even if we get
a failure result, we dont try the internal viewer again.

The only times when internal viewer is tried in this flow is when
either a exception occurs when trying the external viewer, and or
when mime type of the uri is not known.