Viewer: Visibility control, setResult for Video Error
As the viewer activity contains both WebView as well as VideoView,
one requires to keep both disabled and invisible by default and
inturn enable one of them based on what content is to be shown.

TO achieve the same visibility of the view is controlled now.

Also if a error is identified during video playback, then result
code for the activity is set to canceled (i.e error).