Viewer:Video: MediaController, true Error, handleCOntent
Moved logic which calls WebV or VideoV as required into handleContent.
In future other content views may also be added.

Cleaned up some of the logs.

Now Video OnErrorListener doesnt propagate the error, it notifies
completion of error handling to others. This avoids popup of the
default error notification dialog.

Added a MediaController to the VideoView.