SimpRecycView:P1: An attempt at a simple-all-in RecyclerView
For simple situations, where there is a simple String array that we
want to show to user in a efficient and easy way, trying to create
this all-in-one recycler view, which includes the ViewAdapter and
the ViewHolder within itself, so that there is no need to define
them seperately, unlike the regular RecyclerView.

As I am new to kotlin and havent looked at intricacies of android's
gui framework for a long time now, so not sure things are fine yet,
nor have I looked into the documents, just giving it a try.

THis is the initial skeleton.

There are some kotlin mechanisms that I am not sure yet, like why

* I require to specifiy the dataList as val/var explicit as part of
the class definition, for me to acces from its own member functions.

* Or why I cant access the dataList in the outer class directly.
May be I require to use a special tag like outer or so, no idea yet.