SimpRecycView: Avoid ViewAdapter dataList duplication
As noted in my last commit, the need for val in class definition
remained somewhere back in head and as expected it told why not
apply the same to the top level class itself and so made dataList
of the SimpRecycView itself val explicitly in the class + constructor
definition and now I am able to access the same from the inner classes

NOTE: However I am not sure this or even previous definition where
this dataList is part of the constructor is fine. Because when the
class is being inflated from the view, how will one assing the list,
which is instanciated at runtime. So I think I will have to remove
it from class+constructor definition and add a seperate helper func
to assign it later.