SimpRecycViewer:P3: Avoid dataList in default constructor
As I had noted in my last and previous commits, I haven't spent time
yet to look into Android GUI framework flow recently, so for now I
am keeping things simple and avoiding passing the dataList that will
be shown to the user, through the default constructor.

Instead now a seperate function is added, which will have to be called
and the dataList assinged as required. This also allows the dataList
to be changed if required during the life of the app.