SimpRecycView:P5: Secondary constructor
As inflate from gui xml definition expects a constructor which
takes the context as well as the attributeset interface, so have
added a secondary constructure.

NOTE: Realised the reason why one had to define class definition
argument as val or val to access it from the class or its inner
classes. The reason is that rather the class definition also
potentially defines the primary constructor and inturn the arguments
are the arguments into this primary constructor, so they cant be
accessed from outside the init block (I assume) or primary constructor.

But by also specifying a val or var before the argument, we are telling
that we want to make it a property in the class. And so inturn we
can access it from within anywhere in the class or its inner classes.

TODO: The program runs properly without any crash or so, but the view
is still not visible, have to check why still.