SImpRecycView: Explicit constructors
As I couldn't seem to get to call both the primary constructor and
the base classes constructor at the same time as part of the 2ndary
constructor definition (after :). Note: Even thou doc seems to state
the 2ndary constructor implicitly calls primary constructor, the
compiler was giving error without me explicitly adding this(arg) to
denote the primary constructor as part of the 2ndary constructor
definition/declaration.

So now I have changed to having explicit constructors and not have
the primary constructor defined as part of the class name definition.

I am assuming the first constructor is treated as the primary constructor
but I am not sure.

TODO: Doesnt seem like 1st constr is the primary constructor, nothing is
visible.