SimpRecycView:P8: Focusable(Scrollable), Highlight+
TO allow the user to scroll through the list if it is longer than
the screen visible area, the text view has been made focusable.

Have also added support for highlighting the list items, by assigning
a custom hardcoded drawable as the background for now. For now I am
using the background drawable which I already defined for the main
activities recyclerview here also. Need to make this configurable
by the user through the properties in the layout xml file. I assume
this is where that attribute set comes into picture.

Inturn as the view by default wraps content, so have changed to
match parent along width, so that the highlight is properly visible
and clean.