SimpRecycView:P9: Font, color, and failed vertical align
Used monospace font to make the content displayed appear cleaner.

Same with color, otherwise it was using a light gray color, which
wasnt legible fully. So now black is used.

Tried setting the gravity to center_vertical, but it doesnt seem
to work, if one is using wrap content for height.