ArchiveMa: Use BufferedInputStream to support mark
As ArchiveStreamFactory detect logic depends on mark being supported
by the stream passed to it, so on thinking further, passing buffered
inputstream, should provide the same (what was I thinking yesterday ;-)

With it there is no longer the need to use the file extension to archive
type mapper logic.

May explicitly call detect, if required in future, similar to what I
had initially started out with.